feat(dependencies): update package versions and clean up Cargo.lock
- Upgraded `pingora`, `pingora-core`, and `pingora-proxy` to version 0.7.0, enhancing the deployment platform's capabilities.
- Removed outdated dependencies such as `atty` and `clap` version 3.2.25, replacing them with `clap` version 4.5.50, which no longer relies on `atty`.
- Updated `lru` from version 0.14.0 to 0.16.3, ensuring improved performance and compatibility.
- Cleaned up `Cargo.lock` by removing unnecessary package entries and ensuring all dependencies are up-to-date.

This commit significantly improves the dependency management and overall stability of the project.

+++ crates/temps-query/src/types.rs
@@ -223,7 +223,7 @@ impl QueryResult {
             rows,
             stats: QueryStats {
                 row_count,
-                total_rows: Some(row_count as usize),
+                total_rows: Some(row_count),
                 execution_ms,
                 has_more,
                 next_cursor: None,
